# libct/cg/devices: stop using regex