Looking into data generated by setting

	GODEBUG="inittrace=1"

I have noticed this line:

	init github.com/opencontainers/runc/libcontainer/cgroups/devices @1.2 ms, 0.020 ms clock, 10512 bytes, 133 allocs

This is the leader for both bytes and allocs among the packages from
this repo, and all of it is caused by a single regex:

	var devicesListRegexp = regexp.MustCompile(`^([abc])\s+(\d+|\*):(\d+|\*)\s+([rwm]+)$`)

It seems that the same parsing can be done without relying on
a regular expression, no decrease in readability or correctness
(including catching all invalid input), and 2x faster (according to
the benchmark added), and also hopefully makes runc start
slightly faster and leaner.

Before:

	BenchmarkParseLine-4   	  176240	      6768 ns/op	    6576 B/op	      64 allocs/op

After:

	BenchmarkParseLine-4   	  322441	      3535 ns/op	    5520 B/op	      53 allocs/op

[v2: single split with SplitFunc; fix a typo in an error message]
[v3: rebase after #3159 merge; re-ran benchmarks (results are similar)]