@kolyshkin 2021-08-24
Rebased after #3159 merge; PTAL @thaJeztah @cyphar